requirements.txt
flask
openai
python-dotenv
gunicorn
pathlib
ffmpeg
pymupdf